refactor: move escape method to util class
