+++ dev/com.ibm.ws.jndi.iiop/src/com/ibm/ws/jndi/iiop/OrbContext.java
@@ -12,6 +12,8 @@
  *******************************************************************************/
 package com.ibm.ws.jndi.iiop;
 
+import static com.ibm.ws.jndi.iiop.CosNameUtil.escapeCorbanameUrlIfNecessary;
+
 import java.util.Hashtable;
 
 import javax.naming.Binding;
@@ -50,7 +52,7 @@ static final TraceComponent tc = Tr.register(OrbContext.class);
     @FFDCIgnore(NO_IMPLEMENT.class)
     public Object lookup(String name) throws NamingException {
         final String methodName = "lookup(): ";
-        name = CorbanameUrlContextFactory.Escaper.escapeCorbanameUrlIfNecessary(name);
+        name = escapeCorbanameUrlIfNecessary(name);
         if (tc.isDebugEnabled()) Tr.debug(tc, methodName + "");
         org.omg.CORBA.Object result = orb.string_to_object(name);
         if (tc.isDebugEnabled()) Tr.debug(tc, methodName + ": orb.string_to_object() returned "+result);
