+++ dev/com.ibm.ws.jndi.iiop/test/com/ibm/ws/jndi/iiop/CorbanameEscapeTestSuite.java
@@ -1,16 +1,18 @@
 /* ***************************************************************************
- * Copyright (c) 2015 IBM Corporation and others.
+ * Copyright (c) 2015,2024 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
  *     IBM Corporation - initial API and implementation
  * ***************************************************************************/
 package com.ibm.ws.jndi.iiop;
+
+import static com.ibm.ws.jndi.iiop.CosNameUtil.escapeCorbanameUrlIfNecessary;
 
 import java.util.Arrays;
 import java.util.List;
@@ -213,7 +215,7 @@ public class CorbanameEscapeTestSuite {
     }
 
     private static void assertEscaped(String from, String to) {
-        String actual = TestFacade.escapeCorbanameUrlIfNecessary(from);
+        String actual = escapeCorbanameUrlIfNecessary(from);
         Assert.assertEquals(to, actual);
     }
 }
