+++ dev/com.ibm.ws.jndi.iiop/test/com/ibm/ws/jndi/iiop/TestFacade.java
@@ -1,10 +1,10 @@
 /* ***************************************************************************
- * Copyright (c) 2015 IBM Corporation and others.
+ * Copyright (c) 2015,2024 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
@@ -89,10 +89,6 @@ public class TestFacade {
             result[i] = new NameComponent(id, kind);
         }
         return result;
-    }
-
-    public static String escapeCorbanameUrlIfNecessary(String url) {
-        return CorbanameUrlContextFactory.Escaper.escapeCorbanameUrlIfNecessary(url);
     }
 
     private static class TestFacadeDelegate {
